Add files via upload
A 2D platformer I made using Greenfoot in early 2019. It puts the player into the role of a member of the Welsh terrorist group Mudiad Amdiffyn Cymru(movement for the defence of Wales). They were active in the 60s and constantly made threats to destroy structures such as the Severn bridge(now called Prince of Wales bridge) but only attacked on 2 occasions; one was an assassination attempt on Prince Charles where they planned to plant a bomb on the railway he was travelling by, unfortunately for them the bomb blew up early  and killed all participating members. Fortunately in this game, blowing yourself up is a key mechanic and is useful for reaching greater heights. (there are only 4 levels or so)