Add files via upload
Made in Greenfoot as my first step towards making a functional game which I planned to make in Unity after figuring out some things in Greenfoot first. This is essentially AWE 0, with AWE 1 being the somewhat unsuccessful Unity port of this game and AWE 2 being a whole new 3D version which I'm still developing(it's very much a WIP).